refactor(client): shared scroll controller for channel and DM lists
The channel scroll controller and the DM's hand-rolled duplicate merge
into one chat-primitives hook (position memory, load-more prepend
compensation, near-bottom auto-follow, the four-attempt restore ladder,
and the channel-only detached-window paging, jump-latch ownership, and
new-messages-divider landing). The channel keeps a thin adapter so its
call site is unchanged; the DM's inline scroll block is deleted and
calls the hook with dm:-prefixed keys.

Channels keep their saved positions (already server:channel keyed); DM
scroll positions reset once on upgrade (their old bare-number keys are
dropped by the colon-key loader) — acceptable in alpha. Final tranche
of the chat-stack deduplication.
